update tests to follow better ava practices
