--- packages/knack-consumer-client/test/utils/index.js
@@ -1,7 +1,0 @@
-const silentTestLogger = require('./silentTestLogger');
-const knackMocker = require('./knackMocker');
-
-module.exports = {
-	silentTestLogger,
-	knackMocker
-};
