--- packages/knack-consumer-client/test/utils/silentTestLogger.js
@@ -1,8 +1,0 @@
-module.exports = {
-	trace: () => {},
-	debug: () => {},
-	info: () => {},
-	warn: () => {},
-	error: () => {},
-	fatal: () => {}
-};
