+++ packages/knack-consumer/test/cdWithDefaults.test.js
@@ -1,13 +1,55 @@
 const test = require('ava');
+const {EventEmitter} = require('events');
+const {spy} = require('sinon');
 const proxyquire = require('proxyquire');
-const {silentTestLogger, nodeRdMocker} = require('./utils');
+
+const silentTestLogger = {
+	trace: () => {},
+	debug: () => {},
+	info: () => {},
+	warn: () => {},
+	error: () => {},
+	fatal: () => {}
+};
+
+function buildConsumerMock() {
+	class ConsumerMock extends EventEmitter {
+		constructor(consumerConfig, topicConfig) {
+			super();
+			this.consumerConfig = consumerConfig;
+			this.topicConfig = topicConfig;
+			this.connectEventName = 'ready';
+			this.disconnectEventName = 'disconnected';
+			this.consume = spy(this.consume.bind(this));
+			this.connect = spy(this.connect.bind(this));
+			this.disconnect = spy(this.disconnect.bind(this));
+			this.subscribe = spy(this.subscribe.bind(this));
+		}
+
+		consume() {}
+
+		connect() {
+			this.emit(this.connectEventName);
+		}
+
+		disconnect() {
+			this.emit(this.disconnectEventName);
+		}
+
+		subscribe() {}
+	}
+
+	Object.setPrototypeOf(ConsumerMock, spy());
+
+	return ConsumerMock;
+}
 
 test.before('setup test', t => {
 	proxyquire
 		.noCallThru()
 		.noPreserveCache();
 
-	const KafkaConsumer = nodeRdMocker.buildConsumerMock();
+	const KafkaConsumer = buildConsumerMock();
 
 	t.context.knackConsumer = proxyquire('..', {
 		'node-rdkafka': {
