--- packages/knack-consumer/test/utils/index.js
@@ -1,7 +1,0 @@
-const silentTestLogger = require('./silentTestLogger');
-const nodeRdMocker = require('./nodeRdMocker');
-
-module.exports = {
-	silentTestLogger,
-	nodeRdMocker
-};
