+++ packages/knack-producer-client/package.json
@@ -18,9 +18,9 @@
   "publishConfig": {
     "access": "public"
   },
-	"engines": {
-		"node": ">=10.18.0 <11 || >=12.14.0 <13 || >=13.5.0"
-	},
+  "engines": {
+    "node": ">=10.18.0 <11 || >=12.14.0 <13 || >=13.5.0"
+  },
   "scripts": {
     "test": "xo && nyc ava"
   },
@@ -51,19 +51,12 @@
     ]
   },
   "ava": {
-    "files": [
-			"!test/utils/**/*.js"
-		],
-    "sources": [
-      "index.js"
-    ],
     "concurrency": 5,
     "failFast": true,
     "failWithoutAssertions": false,
     "environmentVariables": {},
     "tap": false,
-    "verbose": true,
-    "compileEnhancements": false
+    "verbose": true
   },
   "xo": {
     "rules": {
